Extract Correspondence routing controller
The correspondence routing page will be managed by its
own correspondence routing controller. So we can cleanly
use #new and #create actions for display the routing options
and execute the redirections on request.
