+++ psd-web/app/forms/correspondence_routing_form.rb
@@ -1,4 +1,4 @@
-class CorrespondenceForm
+class CorrespondenceRoutingForm
   include ActiveModel::Model
   include ActiveModel::Attributes
