+++ psd-web/config/locales/en.yml
@@ -137,7 +137,7 @@ en:
             owner_id:
               blank: Select case owner
               not_found: User or team not found
-        correspondence_form:
+        correspondence_routing_form:
           attributes:
             type:
               blank: "Correspondence type must not be empty"
